chore: move ui files to separate folder and fix file names
Also:
- adjust yml file to run ui specs in separate folder from api run
